test: refactor the fixtures in the manager UT
