+++ tests/study/business/conftest.py
@@ -1,0 +1,21 @@
+import contextlib
+
+import pytest
+from antarest.dbmodel import Base
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+
+
+@pytest.fixture(scope="function", name="db_engine")
+def db_engine_fixture():
+    engine = create_engine("sqlite:///:memory:")
+    Base.metadata.create_all(engine)
+    yield engine
+    engine.dispose()
+
+
+@pytest.fixture(scope="function", name="db_session")
+def db_session_fixture(db_engine):
+    make_session = sessionmaker(bind=db_engine)
+    with contextlib.closing(make_session()) as session:
+        yield session
